requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
defusedxml==0.7.1
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jira==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
pyee==9.0.4
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.0
slack==0.0.2
slackeventsapi==3.0.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
